Изменен модуль Role

+++ backendApi/src/modules/roles/roles.controller.ts
@@ -1,10 +1,9 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Put, UseGuards, Req } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Put, UseGuards, Req, HttpException, HttpStatus } from '@nestjs/common';
 import { RolesService } from './roles.service';
-import { CreateRoleDto } from './dto/create-role.dto';
-import { UpdateRoleDto } from './dto/update-role.dto';
+import { CreateRoleDto, UpdateRoleDto } from './dto';
 import { ApiBearerAuth, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
-import { Role } from './entities/role.entity';
 import { JwtAuthGuard } from 'src/modules/auth/guards/auth.guard';
+import { AppError } from 'src/common/constants/error';
 
 @ApiBearerAuth()
 @ApiTags('roles')
@@ -16,13 +15,13 @@ export class RolesController {
   @Post()
   @ApiOperation({ summary: 'Создание роли' })
   @ApiResponse({ status: 201, description: 'Роль успешно создана!' })
-  create(@Body() createRoleDto: CreateRoleDto, @Req() request) {
+  async create(@Body() createRoleDto: CreateRoleDto, @Req() request) {
     return this.rolesService.create(createRoleDto, request.user.user_id);
   }
 
   @Get('all')
   @ApiOperation({ summary: 'Получение списка ролей' })
-  findAll() {
+  async findAll() {
     return this.rolesService.findAll();
   }
 
@@ -31,7 +30,16 @@ export class RolesController {
   @ApiOperation({ summary: 'Изменение отдельной роли' })
   @ApiResponse({ status: 200, description: 'Роль успешно обновлена!' })
   @ApiResponse({ status: 404, description: 'Роль не существует!' })
-  update(@Body() updateRoleDto: UpdateRoleDto, @Req() request) {
+  async update(@Body() updateRoleDto: UpdateRoleDto, @Req() request) {
+    let foundRole = null;
+    if (updateRoleDto.role_id) {
+      foundRole = await this.rolesService.findOne(updateRoleDto.role_id);
+    }
+
+    if (!foundRole) {
+      throw new HttpException(AppError.ROLE_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+
     return this.rolesService.update(updateRoleDto, request.user.user_id);
   }
 
@@ -40,7 +48,13 @@ export class RolesController {
   @ApiOperation({ summary: 'Удаление отдельной роли' })
   @ApiResponse({ status: 201, description: 'Роль успешно удалена!' })
   @ApiResponse({ status: 404, description: 'Роль не существует!' })
-  remove(@Param('id') id: number, @Req() request) {
+  async remove(@Param('id') id: number, @Req() request) {
+    const foundRole = await this.rolesService.findOne(id);
+
+    if (!foundRole) {
+      throw new HttpException(AppError.ROLE_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+
     return this.rolesService.remove(+id, request.user.user_id);
   }
 }
